Update the Section controller, mapper, service and dao

+++ src/main/java/com/controlcenter/controlcenter/model/Section.java
@@ -12,10 +12,9 @@ public class Section {
     private Long section_id;
     private String section_name;
     private Long dept_id;
-    private Long del_flag;
+    private int del_flag;
     private String reg_id;
     private Timestamp reg_date;
     private String update_id;
     private Timestamp update_date;
-
 }
